Separate centroid rendering process
