+++ js/platform/base.js
@@ -1,7 +1,7 @@
 import ScatterRenderer from '../renderer/scatter.js'
 import LinePlotter from '../renderer/util/lineplot.js'
+import CentroidPlotter from '../renderer/util/centroids.js'
 
-import { DataPointStarPlotter, DataPoint, DataLine } from '../utils.js'
 import TableRenderer from '../renderer/table.js'
 
 export class BasePlatform {
@@ -185,7 +185,10 @@ export class DefaultPlatform extends BasePlatform {
 	init() {
 		this._cur_dimension = this.setting.dimension
 		this.setting.footer.innerText = ''
-		this.svg.querySelector('g.centroids')?.remove()
+		if (this._centroids) {
+			this._centroids.terminate()
+			this._centroids = null
+		}
 		this._renderer.forEach(rend => rend.init())
 		this.render()
 		if (this._loss) {
@@ -196,60 +199,10 @@ export class DefaultPlatform extends BasePlatform {
 	}
 
 	centroids(center, cls, { line = false, duration = 0 } = {}) {
-		let centroidSvg = this.svg.querySelector('g.centroids')
-		if (!centroidSvg) {
-			centroidSvg = document.createElementNS('http://www.w3.org/2000/svg', 'g')
-			centroidSvg.classList.add('centroids')
-			this.svg.appendChild(centroidSvg)
-			const cline = document.createElementNS('http://www.w3.org/2000/svg', 'g')
-			cline.classList.add('c-line')
-			centroidSvg.appendChild(cline)
-			this._centroids_line = []
-			this._centroids = null
+		if (!this._centroids) {
+			this._centroids = new CentroidPlotter(this._renderer[0])
 		}
-		const existCentroids = []
-		if (this._centroids) {
-			this._centroids.forEach(c => {
-				if (Array.isArray(cls) && cls.indexOf(c.category) < 0) {
-					c.remove()
-				} else {
-					existCentroids.push(c)
-				}
-			})
-		}
-		const p = this._renderer[0].points
-		for (let k = 0; k < p.length; k++) {
-			if (this._centroids_line[k]?._from !== p[k] || !line) {
-				this._centroids_line[k]?.remove()
-				this._centroids_line[k] = null
-			}
-		}
-		this._centroids = center.map((c, i) => {
-			let dp = Array.isArray(cls) ? existCentroids.find(e => e.category === cls[i]) : existCentroids[i]
-			if (!dp) {
-				dp = new DataPoint(centroidSvg, this._renderer[0].toPoint(c), Array.isArray(cls) ? cls[i] : cls)
-				dp.plotter(DataPointStarPlotter)
-			}
-			if (line) {
-				const p = this._renderer[0].points
-				const y = this.datas.y
-				for (let k = 0; k < p.length; k++) {
-					if (y[k] === cls[i]) {
-						if (!this._centroids_line[k]) {
-							this._centroids_line[k] = new DataLine(centroidSvg.querySelector('.c-line'), p[k], dp)
-						} else {
-							this._centroids_line[k].to = dp
-						}
-					}
-				}
-			}
-			return dp
-		})
-		Promise.resolve().then(() => {
-			this._centroids.forEach((c, i) => {
-				c.move(this._renderer[0].toPoint(center[i]), duration)
-			})
-		})
+		this._centroids.set(center, cls, { line, duration })
 	}
 
 	_getEvaluateElm() {
